Runners need to be awaited (or waited) before tests.
