+++ tests/NScatterGather.Tests/AggregatedResponseTests.cs
@@ -14,13 +14,13 @@ namespace NScatterGather
 
         public AggregatedResponseTests()
         {
-            _ex = new Exception();
+            _ex = new Exception("Test ex.");
 
             var runner = new RecipientRunner<int>(new Recipient(typeof(object)));
-            runner.Run(_ => Task.FromResult(42));
+            runner.Run(_ => Task.FromResult(42)).Wait();
 
             var runnerFaulted = new RecipientRunner<int>(new Recipient(typeof(bool)));
-            runnerFaulted.Run(_ => Task.FromException<int>(new Exception()));
+            runnerFaulted.Run(_ => Task.FromException<int>(_ex)).Wait();
 
             var runnerIncomplete = new RecipientRunner<int>(new Recipient(typeof(long)));
             runnerIncomplete.Run(_ => GetInfiniteTask<int>());
@@ -68,64 +68,6 @@ namespace NScatterGather
             Assert.Single(completed);
             Assert.Single(faulted);
             Assert.Single(incomplete);
-        }
-    }
-
-    public class AggregatedResponseExtensionsTests
-    {
-        private readonly RecipientRunner<int>[] _runners;
-
-        public AggregatedResponseExtensionsTests()
-        {
-            var runner = new RecipientRunner<int>(new Recipient(typeof(object)));
-            runner.Run(_ => Task.FromResult(42));
-
-            var runnerFaulted = new RecipientRunner<int>(new Recipient(typeof(bool)));
-            runnerFaulted.Run(_ => Task.FromException<int>(new Exception()));
-
-            var runnerIncomplete = new RecipientRunner<int>(new Recipient(typeof(long)));
-            runnerIncomplete.Run(_ => GetInfiniteTask<int>());
-
-            _runners = new[] { runner, runnerFaulted, runnerIncomplete };
-
-            // Local functions.
-
-            static Task<TResult> GetInfiniteTask<TResult>()
-            {
-                var source = new TaskCompletionSource<TResult>();
-                return source.Task;
-            }
-        }
-
-        [Fact]
-        public void Error_if_input_is_null()
-        {
-            Assert.Throws<ArgumentNullException>(() =>
-                (null as AggregatedResponse<int>)!.AsResultsDictionary());
-
-            Assert.Throws<ArgumentNullException>(() =>
-                (null as AggregatedResponse<int>)!.AsResultsList());
-        }
-
-        [Fact]
-        public void Can_be_projected_onto_results_dictionary()
-        {
-            var response = new AggregatedResponse<int>(_runners);
-            var results = response.AsResultsDictionary();
-            Assert.NotNull(results);
-            Assert.Single(results.Keys);
-            Assert.Equal(typeof(object), results.Keys.First());
-            Assert.Single(results.Values);
-            Assert.Equal(42, results.Values.First());
-        }
-
-        [Fact]
-        public void Can_be_projected_onto_results_list()
-        {
-            var response = new AggregatedResponse<int>(_runners);
-            var results = response.AsResultsList();
-            Assert.NotNull(results);
-            Assert.Single(results, 42);
         }
     }
 }
